Converted long returns into reusable components
